Added manifest bashscript
Uses gnu-find to make a list of files.
xargs to pass each file to cksum (SHA256)
the checksum contents are piped to gpg
gpg by signs and encrypts the checksums
and the content is redirected to a file

The default output filename (OFILE) is
.manifest.${cksum}.asc (in the pwd)
The ext matches the --armor output of gpg

To verify the files, the OFILE is
decrypted and the signature checked.
The CKSUM program is called with --quiet
and will echo to STDOUT OK or FAIL
Indicating the status of the check.
Signature status is passed to STDERR;
via gpg default behaviour.